subscription-service/requirements.txt
fastapi>=0.111.0,<0.112.0
uvicorn[standard]>=0.30.0,<0.31.0
uvloop>=0.19.0
httptools>=0.6.1
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
orjson>=3.10.0
sqlalchemy[asyncio]>=2.0.30,<3.0.0
asyncpg>=0.29.0
alembic>=1.13.0,<2.0.0
redis[hiredis]>=5.0.4,<6.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
httpx>=0.27.0
structlog>=24.1.0
